+++ src/components/auth/Register.tsx
@@ -1,9 +1,7 @@
 import { SignedOut, SignUp } from '@clerk/clerk-react'
-import Navbar from '../Landing/Navbar'
 export default function Register() {
   return (
     <>
-      <Navbar/>
     <div className="flex justify-center items-center min-h-screen">
       <SignedOut>
         <SignUp
