+++ src/main.tsx
@@ -3,7 +3,6 @@ import { createRoot } from 'react-dom/client';
 import App from './App.tsx';
 import './index.css';
 import { ClerkProvider} from '@clerk/clerk-react';
-
 const clerkPubKey = import.meta.env.VITE_CLERK_PUBLISHABLE_KEY;
 if (!clerkPubKey) {
   throw new Error('Add your Clerk Publishable Key to the .env.local file')
